Conflicts:
	a2lib/include/TAlphaDisplay.h
	alpha2/src/A2_display_module.cxx
	rootUtils/include/TA2Plot.h

+++ a2lib/build_db.sh
@@ -3,6 +3,8 @@
 cd "${AGRELEASE}/a2lib"
 #Move old database to backup
 DATE=`date +%Y%m%d`
-mv main.db main.${DATE}.db
+if [ -f main.db ]; then
+   mv main.db main.${DATE}.db
+fi
 #Build database
 sqlite3 main.db < settings.sql
